refactor(rust::custom_fees): move tests into separate file
